repo_results_2/run-20240312-153149_LLMatic-ZeroShot-CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-twitter-donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/10/LLMatic/mock_db.py
class MockDB:
	def __init__(self):
		self.data = {}

	def add(self, key, value):
		self.data[key] = value

	def update(self, key, value):
		if key in self.data:
			self.data[key] = value

	def delete(self, key):
		if key in self.data:
			del self.data[key]

	def retrieve(self, key):
		return self.data.get(key, None)

	def add_user(self, email, password):
		self.data[email] = {'password': password, 'blocked_contacts': [], 'groups': {}, 'messages': []}

	def get_user(self, email):
		return self.data.get(email, None)

	def update_user_picture(self, user_id, picture):
		user = self.get_user(user_id)
		if user is not None:
			user['picture'] = picture

	def update_user_status_message(self, user_id, message):
		user = self.get_user(user_id)
		if user is not None:
			user['status_message'] = message

	def update_user_privacy_settings(self, user_id, settings):
		user = self.get_user(user_id)
		if user is not None:
			user['privacy_settings'] = settings

	def block_contact(self, user_id, contact_id):
		user = self.get_user(user_id)
		if user is not None and contact_id not in user['blocked_contacts']:
			user['blocked_contacts'].append(contact_id)

	def unblock_contact(self, user_id, contact_id):
		user = self.get_user(user_id)
		if user is not None and contact_id in user['blocked_contacts']:
			user['blocked_contacts'].remove(contact_id)

	def create_group(self, user_id, group_details):
		user = self.get_user(user_id)
		if user is not None:
			group_id = len(user['groups']) + 1
			group_details['participants'] = [user_id]
			group_details['admins'] = [user_id]
			user['groups'][group_id] = group_details

	def edit_group(self, user_id, group_id, group_details):
		user = self.get_user(user_id)
		if user is not None and group_id in user['groups']:
			user['groups'][group_id].update(group_details)

	def delete_group(self, user_id, group_id):
		user = self.get_user(user_id)
		if user is not None and group_id in user['groups']:
			del user['groups'][group_id]

	def add_participant(self, group_id, participant_id):
		for user in self.data.values():
			if group_id in user['groups']:
				user['groups'][group_id]['participants'].append(participant_id)

	def remove_participant(self, group_id, participant_id):
		for user in self.data.values():
			if group_id in user['groups'] and participant_id in user['groups'][group_id]['participants']:
				user['groups'][group_id]['participants'].remove(participant_id)

	def add_admin(self, group_id, admin_id):
		for user in self.data.values():
			if group_id in user['groups']:
				user['groups'][group_id]['admins'].append(admin_id)

	def remove_admin(self, group_id, admin_id):
		for user in self.data.values():
			if group_id in user['groups'] and admin_id in user['groups'][group_id]['admins']:
				user['groups'][group_id]['admins'].remove(admin_id)

	def send_message(self, sender_id, receiver_id, message):
		sender = self.get_user(sender_id)
		receiver = self.get_user(receiver_id)
		if sender is not None and receiver is not None:
			message_id = len(sender['messages']) + 1
			message_data = {'id': message_id, 'sender_id': sender_id, 'receiver_id': receiver_id, 'message': message, 'read': False}
			sender['messages'].append(message_data)
			receiver['messages'].append(message_data)

	def read_message(self, user_id, message_id):
		user = self.get_user(user_id)
		if user is not None:
			for message in user['messages']:
				if message['id'] == message_id:
					message['read'] = True
					break

	def encrypt_message(self, user_id, message_id, encryption_key):
		user = self.get_user(user_id)
		if user is not None:
			for message in user['messages']:
				if message['id'] == message_id:
					message['message'] = self._encrypt_message(message['message'], encryption_key)
					break

	def _encrypt_message(self, message, encryption_key):
		# This is a mock encryption function. In a real application, you would use a secure encryption algorithm.
		return ''.join(chr((ord(c) + encryption_key) % 256) if c.isprintable() else c for c in message)

	def share_content(self, user_id, receiver_id, content):
		self.send_message(user_id, receiver_id, content)

